backend/ReleaseVault/requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cors==1.0.1
cryptography==36.0.0
dateutil-br==0.0.7
Deprecated==1.2.13
Django==2.2.24
django-cors-headers==2.4.0
django-ranged-response==0.2.0
django-redis==5.0.0
django-simple-captcha==0.5.14
django-simpleui==2023.12.12
django-threadlocals==0.10
djangorestframework==3.13.1
filelock==3.4.0
Flask==2.0.2
future==0.18.2
gevent
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.5
MarkupSafe==2.0.1
mysqlclient==2.1.1
packaging==21.3
Pillow==8.4.0
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.9.4
PyJWT==2.3.0
PyMySQL==0.10.0
pyOpenSSL==19.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.2
tldextract==3.1.2
tzdata==2022.6
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.14.1
xlrd==1.2.0
zope.event==4.5.0
zope.interface==5.4.0
